refactor(docs): ownership-based naming on /challenges
