+++ apps/web/src/routes/challenges/challenges-details-overview-adminbot.svelte
