+++ apps/web/src/routes/challenges/challenges-details-podium-grid.svelte
