+++ apps/web/src/routes/challenges/challenges-details-podium.svelte
@@ -3,7 +3,7 @@
   import { useChallengeSolves, useClientConfig, useCurrentUser } from '$lib/query'
   import type { RankVariant } from '$lib/utils'
   import { formatFirstBloodTime, formatRelativeToFirstBlood, getTimeOrdinal } from '$lib/utils'
-  import ChallengePodiumGrid, { type PodiumItem } from './challenge-podium-grid.svelte'
+  import ChallengePodiumGrid, { type PodiumItem } from './challenges-details-podium-grid.svelte'
 
   interface Props {
     challenge: Challenge
