+++ apps/web/src/routes/challenges/challenges-details-rank-delta.svelte
