+++ apps/web/src/routes/challenges/challenges-details-scores-graph-controls.svelte
