+++ apps/web/src/routes/challenges/challenges-details-scores-graph.svelte
