+++ apps/web/src/routes/challenges/challenges-details-submit.svelte
