+++ apps/web/src/routes/challenges/challenges-list-header.svelte
