+++ apps/web/src/routes/challenges/challenges-list-item.svelte
@@ -2,7 +2,7 @@
   import { ChallengeScoringKind, type Challenge } from '@rctf/types'
   import { IconAwardFilled, IconCheck } from '$lib/icons'
   import { cn, getBloodStyles, type BloodIndex } from '$lib/utils'
-  import ChallengeDynamicDelta from './challenge-dynamic-delta.svelte'
+  import ChallengeDynamicDelta from './challenges-point-delta.svelte'
 
   interface Props {
     challenge: Challenge
