+++ apps/web/src/routes/challenges/challenges-point-delta.svelte
